Add files via upload
Also checks to see if space is available in the column that the user chose. Users input is also checked to make sure there input is within the board and also that the column is not yet full. If the column is full, the user is asked to enter a different column. Once 4 identical flags are connected diagonally, vertically or horizontally. The Console displays that Player X or Player O have won depending on if the flags that connected are X's or O's.